feat!: promote `experimental.alternateLinkCanonicalQueries`
