+++ docs/content/docs/04.api/00.options.md
@@ -474,8 +474,8 @@ This feature relies on [Nuxt's `experimental.typedRoutes`](https://nuxt.com/docs
 ### `alternateLinkCanonicalQueries`
 
 - type: `boolean`{lang="ts-type"}
-- default: `false`{lang="ts"}
-- Whether to remove non-canonical query parameters from alternate link meta tags
+- default: `true`{lang="ts"}
+- Remove non-canonical query parameters from alternate link meta tags
 
 ## `hmr`
