+++ specs/basic_usage.spec.ts
@@ -16,10 +16,7 @@ describe('basic usage', async () => {
           i18n: {
             baseUrl: 'http://localhost:3000',
             skipSettingLocaleOnNavigate: undefined,
-            detectBrowserLanguage: undefined,
-            experimental: {
-              alternateLinkCanonicalQueries: false
-            }
+            detectBrowserLanguage: undefined
           }
         }
       }
