+++ specs/fixtures/basic_usage/nuxt.config.ts
@@ -18,9 +18,6 @@ export default defineNuxtConfig({
     baseUrl: 'http://localhost:3000',
     vueI18n: './config/i18n.config.ts',
     locales: ['en', 'fr'],
-    defaultLocale: 'en',
-    experimental: {
-      alternateLinkCanonicalQueries: false
-    }
+    defaultLocale: 'en'
   }
 })
