+++ specs/fixtures/basic_usage_compat_4/nuxt.config.ts
@@ -23,7 +23,6 @@ export default defineNuxtConfig({
     locales: ['en', 'fr'],
     defaultLocale: 'en',
     experimental: {
-      alternateLinkCanonicalQueries: false,
       localeDetector: './localeDetector.ts'
     }
   }
